paper1-latent-str/requirements-lock.txt
# Deep-LSTR: Locked Dependencies for Reproducibility
# Generated for Python 3.12, PyTorch 2.6 environment
# Last verified: 2025-01-17

# Core scientific computing
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# Deep learning
torch==2.6.0

# Data acquisition
yfinance==0.2.48

# Parallel processing
joblib==1.4.2

# Visualization
matplotlib==3.9.2

# Progress bars
tqdm==4.66.5

# YAML configuration
pyyaml==6.0.2

# Transitive dependencies (key packages)
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.4
fsspec==2024.9.0
html5lib==1.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
multitasking==0.0.11
networkx==3.4.1
packaging==24.1
peewee==3.17.7
pillow==10.4.0
platformdirs==4.3.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sympy==1.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
